requirements.txt
fastapi~=0.110.2
joblib==1.2.0
pandas~=2.1.4
pendulum==3.0.0
pydantic~=2.5.3
Requests==2.31.0
SQLAlchemy~=2.0.29
streamlit==1.32.0
xgboost==2.0.3
apache-airflow==2.9.0
scikit-learn~=1.3.0
